feat(web): integrate Memory Map into the shared sidebar layout
/memory already used the shared ChatLayout.vue shell, but
ProjectSidebar.vue had no template branch for mode === 'memory', so it
rendered an empty column while MemoryMapView.vue drew its own separate
sidebar (vault stats/search/categories/path finder) squeezed next to
it. Follows the same shared-store split already used for schedules
(useTaskStore): a new memoryMap.ts Pinia store holds the graph data and
filter state, ProjectSidebar renders the workspace toggle plus the
moved-in vault controls, and MemoryMapView keeps only what's page-
specific (canvas rendering, list-view sort).

Fixes two bugs found while doing this:
- The graph going blank on workspace switch: the <canvas> element is
  destroyed and recreated on every reload (the v-if="loading" branch),
  but the 2D context and ResizeObserver were only ever captured once in
  onMounted, so after any reload they pointed at a detached canvas.
  Re-attaches on every (re)appearance instead.
- The browser tab title falling through to a stale active-chat title
  on this page (ChatLayout.vue's pageDocumentTitle had no memory-mode
  branch).

The 1-9 keyboard workspace shortcut already worked; now that the graph
store watches the same shared activeWorkspace, it reloads correctly
through that path too (verified by reproduction).

vue-tsc clean, 652/652 frontend tests pass, npm run build succeeds.
No new unit tests were added for the store/bug fixes; verified via
Playwright browser testing instead.
